backend/requirements.txt
google-adk
fastapi
uvicorn[standard]
httpx
python-dotenv
pydantic
